database/requirements.txt
pydantic
pyhumps
sqlalchemy
sqlite3
